Update doc on how to sync_host --> configure and start zookeeper --> configure hadoop

+++ hadoop/repo/hadoop/download_util.sh
@@ -15,5 +15,10 @@ FULL_DESTINATION_PATH="$DESTINATION_DIR/$FILENAME"
 # Create the destination directory if it doesn't exist
 mkdir -p "$DESTINATION_DIR"
 
-# Download the file with wget, showing a progress bar
-wget -O "$FULL_DESTINATION_PATH" --progress=bar:force "$URL"
+# Check if the file already exists
+if [ ! -f "$FULL_DESTINATION_PATH" ]; then
+    # Download the file with wget, showing a progress bar
+    wget -O "$FULL_DESTINATION_PATH" --progress=bar:force "$URL"
+else
+    echo "File '$FULL_DESTINATION_PATH' already exists. Download skipped."
+fi
